cleanup free callback (this saved 2k bytes lol)
