+++ alloc/alloc.h
@@ -3,5 +3,7 @@
 FlipWorldApp *flip_world_app_alloc();
 void flip_world_app_free(FlipWorldApp *app);
 void flip_world_show_submenu();
-bool alloc_view_loader(void *context);
-void free_view_loader(void *context);
+uint32_t callback_exit_app(void *context);
+uint32_t callback_to_submenu(void *context);
+uint32_t callback_to_wifi_settings(void *context);
+uint32_t callback_to_settings(void *context);
