+++ callback/loader.h
@@ -39,10 +39,10 @@ struct DataLoaderModel
     FuriTimer *timer;
     FlipperHTTP *fhttp;
 };
+bool loader_view_alloc(void *context);
+void loader_view_free(void *context);
 void loader_switch_to_view(FlipWorldApp *app, char *title, DataLoaderFetch fetcher, DataLoaderParser parser, size_t request_count, ViewNavigationCallback back, uint32_t view_id);
-
 void loader_draw_callback(Canvas *canvas, void *model);
-
 void loader_init(View *view);
 void loader_widget_set_text(char *message, Widget **widget);
 void loader_free_model(View *view);
